Split Session and Image into different files. Fix the documentation.
