+++ carta/session.py
@@ -1,41 +1,38 @@
 """
-This is the main module of the CARTA Python wrapper. It comprises a session class which represents a CARTA frontend session, and an image class which represents a single image open in the session.
+This is the main module of the CARTA Python wrapper. It contains a session class which represents a CARTA frontend session.
 
-The user can interact with an existing CARTA session open in their browser by creating a session object using the :obj:`carta.client.Session.interact` classmethod.
+The user can interact with an existing CARTA session open in their browser by creating a session object using the :obj:`carta.session.Session.interact` classmethod.
 
-Alternatively, the user can create a new session which runs in a headless browser controlled by the wrapper. The user can connect to an existing CARTA backend instance (using the :obj:`carta.client.Session.connect` classmethod), or first start a new CARTA backend instance which is also controlled by the wrapper (using the :obj:`carta.client.Session.new` classmethod). The backend can be started either on the local host or on a remote host which the user can access with passwordless SSH.
-
-Image objects should not be instantiated directly, and should only be created through methods on the session object.
+Alternatively, the user can create a new session which runs in a headless browser controlled by the wrapper. The user can connect to an existing CARTA backend instance (using the :obj:`carta.session.Session.connect` classmethod), or first start a new CARTA backend instance which is also controlled by the wrapper (using the :obj:`carta.session.Session.new` classmethod). The backend can be started either on the local host or on a remote host which the user can access with passwordless SSH.
 """
 
-import posixpath
 import base64
 
-from .constants import Colormap, Scaling, CoordinateSystem, LabelType, BeamType, PaletteColor, Overlay, SmoothingMode, ContourDashMode
+from .image import Image
+from .constants import CoordinateSystem, LabelType, BeamType, PaletteColor, Overlay
 from .protocol import Protocol
-from .util import logger, Macro, cached, split_action_path, CartaScriptingException
-from .validation import validate, String, Number, Color, Constant, Boolean, NoneOr, IterableOf, OneOf, Evaluate, Attr
+from .util import logger, Macro, split_action_path, CartaScriptingException
+from .validation import validate, String, Number, Color, Constant, Boolean, NoneOr, OneOf
 
-# TODO split Session and Image into two files!
 class Session:
     """This object corresponds to a CARTA frontend session.
     
     This class provides the core generic method for calling actions on the frontend (through the backend), as well as convenience methods which wrap this generic method and provide a more intuitive and user-friendly interface to frontend functionality associated with the session as a whole.
     
-    This class should not be instantiated directly. Three class methods are provided for creating different types of sessions with all the appropriate parameters set: :obj:`carta.client.Session.interact` for interacting with an existing CARTA session open in the user's browser, :obj:`carta.client.Session.create` for creating a new CARTA session in a headless browser by connecting to an existing CARTA backend or controller instance, and :obj:`carta.client.Session.start_and_create` for starting a new backend instance and then connecting to it to create a new session in a headless browser.
+    This class should not be instantiated directly. Three class methods are provided for creating different types of sessions with all the appropriate parameters set: :obj:`carta.session.Session.interact` for interacting with an existing CARTA session open in the user's browser, :obj:`carta.session.Session.create` for creating a new CARTA session in a headless browser by connecting to an existing CARTA backend or controller instance, and :obj:`carta.session.Session.start_and_create` for starting a new backend instance and then connecting to it to create a new session in a headless browser.
     
     The session object can be used to create image objects, which provide analogous convenience methods for functionality associated with individual images.
     
     Parameters
     ----------
     session_id : integer
-        The ID of the CARTA frontend session associated with this object. This is set automatically when a new session is created with :obj:`carta.client.Session.create` or :obj:`carta.client.Session.start_and_create`.
+        The ID of the CARTA frontend session associated with this object. This is set automatically when a new session is created with :obj:`carta.session.Session.create` or :obj:`carta.session.Session.start_and_create`.
     protocol : :obj:`carta.protocol.Protocol`
         The protocol object which handles HTTP communication with the CARTA scripting API. This is created automatically when a new session is created with one of the three class methods for creating sessions.
     browser : :obj:`carta.browser.Browser`
-        The browser object associated with this session. This is created automatically when a new session is created with :obj:`carta.client.Session.create` or :obj:`carta.client.Session.start_and_create`.
+        The browser object associated with this session. This is created automatically when a new session is created with :obj:`carta.session.Session.create` or :obj:`carta.session.Session.start_and_create`.
     backend : :obj:`carta.browser.Backend`
-        The backend object associated with this session. This is created automatically when a new session is created with :obj:`carta.client.Session.start_and_create`.
+        The backend object associated with this session. This is created automatically when a new session is created with :obj:`carta.session.Session.start_and_create`.
     debug_no_auth : boolean
         This should be set if the backend has been started with the ``--debug_no_auth`` option. This is provided for debugging purposes only and should not be used under normal circumstances.
     """
@@ -68,7 +65,7 @@ class Session:
             
         Returns
         -------
-        :obj:`carta.client.Session`
+        :obj:`carta.session.Session`
             A session object associated with the frontend session provided.
             
         Raises
@@ -101,7 +98,7 @@ class Session:
             
         Returns
         -------
-        :obj:`carta.client.Session`
+        :obj:`carta.session.Session`
             A session object connected to a new frontend session running in the browser provided.
             
         Raises
@@ -136,7 +133,7 @@ class Session:
             
         Returns
         -------
-        :obj:`carta.client.Session`
+        :obj:`carta.session.Session`
             A session object connected to a new frontend session running in the browser provided.
             
         Raises
@@ -188,7 +185,7 @@ class Session:
     def get_value(self, path):
         """Get the value of an attribute from a frontend store.
         
-        Like the :obj:`carta.client.Session.call_action` method, this is exposed in the public API but is not intended to be used directly under normal circumstances.
+        Like the :obj:`carta.session.Session.call_action` method, this is exposed in the public API but is not intended to be used directly under normal circumstances.
         
         Parameters
         ----------
@@ -260,7 +257,7 @@ class Session:
         
         TODO: .. is not supported, but it can be now that we have made this value independent of the frontend.
         
-        This does not affect the starting directory saved by the frontend. To change that directory, use :obj:`carta.client.session.set_starting_directory`.
+        This does not affect the starting directory saved by the frontend. To change that directory, use :obj:`carta.session.Session.set_starting_directory`.
         
         Parameters
         ----------
@@ -276,7 +273,7 @@ class Session:
         
         It should not be necessary to use this method in non-interactive scripts which do not reuse frontend sessions.
         
-        This does not affect the current directory used by the wrapper. To change that directory, use :obj:`carta.client.session.cd`.
+        This does not affect the current directory used by the wrapper. To change that directory, use :obj:`carta.session.Session.cd`.
         
         This method must be called before any methods that use the locally saved path.
         
@@ -321,7 +318,7 @@ class Session:
         
         Returns
         -------
-        list of :obj:`carta.client.Image` objects.
+        list of :obj:`carta.image.Image` objects.
         """
         return Image.from_list(self, self.get_value("frameNames"))
     
@@ -330,7 +327,7 @@ class Session:
         
         Returns
         -------
-        :obj:`carta.client.Image`
+        :obj:`carta.image.Image`
             The currently active image.
         """
         frame_info = self.get_value("activeFrame.frameInfo")
@@ -614,458 +611,3 @@ class Session:
             
         if self._backend is not None:
             self._backend.stop()
-
-
-class Image:
-    """This object corresponds to an image open in a CARTA frontend session.
-    
-    This class should not be instantiated directly. Instead, use the session object's methods for opening new images or retrieving existing images.
-    
-    Parameters
-    ----------
-    session : :obj:`carta.client.Session`
-        The session object associated with this image.
-    image_id : integer
-        The ID identifying this image within the session. This is a unique number which is not reused, not the index of the image within the list of currently open images.
-    file_name : string
-        The file name of the image. This is not a full path.
-    
-    Attributes
-    ----------
-    session : :obj:`carta.client.Session`
-        The session object associated with this image.
-    image_id : integer
-        The ID identifying this image within the session.
-    file_name : string
-        The file name of the image.
-    
-    """
-    def __init__(self, session, image_id, file_name):
-        self.session = session
-        self.image_id = image_id
-        self.file_name = file_name
-        
-        self._base_path = f"frameMap[{image_id}]"
-        self._frame = Macro("", self._base_path)
-    
-    @classmethod
-    def new(cls, session, path, hdu, append):
-        """Open or append a new image in the session and return an image object associated with it.
-        
-        This method should not be used directly. It is wrapped by :obj:`carta.client.Session.open_image` and :obj:`carta.client.Session.append_image`.
-        
-        Parameters
-        ----------
-        session : :obj:`carta.client.Session`
-            The session object.
-        path : string
-            The path to the image file, either relative to the session's current directory or an absolute path relative to the CARTA backend's root directory.
-        hdu : string
-            The HDU to open.
-        append : boolean
-            Whether the image should be appended. By default it is not, and all other open images are closed.
-        
-        Returns
-        -------
-        :obj:`carta.client.Image`
-            A new image object.
-        """
-        path = session.resolve_file_path(path)
-        directory, file_name = posixpath.split(path)
-        image_id = session.call_action("appendFile" if append else "openFile", directory, file_name, hdu, return_path="frameInfo.fileId")
-        
-        return cls(session, image_id, file_name)
-    
-    @classmethod
-    def from_list(cls, session, image_list):
-        """Create a list of image objects from a list of open images retrieved from the frontend.
-        
-        This method should not be used directly. It is wrapped by :obj:`carta.client.Session.image_list`.
-        
-        Parameters
-        ----------
-        session : :obj:`carta.client.Session`
-            The session object.
-        image_list : list of dicts
-            The JSON object representing frame names retrieved from the frontend.
-            
-        Returns
-        -------
-        list of :obj:`carta.client.Image`
-            A list of new image objects.
-        """
-        return [cls(session, f["value"], f["label"].split(":")[1].strip()) for f in image_list]
-        
-    def __repr__(self):
-        return f"{self.session.session_id}:{self.image_id}:{self.file_name}"
-    
-    def call_action(self, path, *args, **kwargs):
-        """Convenience wrapper for the session object's generic action method.
-        
-        This method calls :obj:`carta.client.Session.call_action` after prepending this image's base path to the path parameter.
-        
-        Parameters
-        ----------
-        path : string
-            The path to an action relative to this image's frame store.
-        *args
-            A variable-length list of parameters. These are passed unmodified to the session method.
-        **kwargs
-            Arbitrary keyword parameters. These are passed unmodified to the session method.
-        
-        Returns
-        -------
-            object or None
-                The unmodified return value of the session method.
-        """
-        return self.session.call_action(f"{self._base_path}.{path}", *args, **kwargs)
-    
-    def get_value(self, path):
-        """Convenience wrapper for the session object's generic method for retrieving attribute values.
-        
-        This method calls :obj:`carta.client.Session.get_value` after prepending this image's base path to the path parameter.
-        
-        Parameters
-        ----------
-        path : string
-            The path to an attribute relative to this image's frame store.
-        
-        Returns
-        -------
-            object
-                The unmodified return value of the session method.
-        """
-        return self.session.get_value(f"{self._base_path}.{path}")
-    
-    # METADATA
-    
-    @property
-    @cached
-    def directory(self):
-        """The path to the directory containing the image."""
-        return self.get_value("frameInfo.directory")
-    
-    @property
-    @cached
-    def header(self):
-        """The header of the image."""
-        return self.get_value("frameInfo.fileInfoExtended.headerEntries")
-    
-    @property
-    @cached
-    def shape(self):
-        """The shape of the image."""
-        return list(reversed([self.width, self.height, self.depth, self.stokes][:self.ndim]))
-    
-    @property
-    @cached
-    def width(self):
-        """The width of the image."""
-        return self.get_value("frameInfo.fileInfoExtended.width")
-    
-    @property
-    @cached
-    def height(self):
-        """The height of the image."""
-        return self.get_value("frameInfo.fileInfoExtended.height")
-    
-    @property
-    @cached
-    def depth(self):
-        """The depth of the image."""
-        return self.get_value("frameInfo.fileInfoExtended.depth")
-    
-    @property
-    @cached
-    def stokes(self):
-        """The number of Stokes parameters of the image."""
-        return self.get_value("frameInfo.fileInfoExtended.stokes")
-    
-    @property
-    @cached
-    def ndim(self):
-        """The number of dimensions of the image."""
-        return self.get_value("frameInfo.fileInfoExtended.dimensions")
-    
-    # SELECTION
-    
-    def make_active(self):
-        """Make this the active image."""
-        self.session.call_action("setActiveFrameById", self.image_id)
-        
-    def make_spatial_reference(self):
-        """Make this image the spatial reference."""
-        self.session.call_action("setSpatialReference", self._frame)
-        
-    @validate(Boolean())
-    def set_spatial_matching(self, state):
-        """Enable or disable spatial matching.
-        
-        Parameters
-        ----------
-        state : boolean
-            The desired spatial matching state.
-        """
-        self.session.call_action("setSpatialMatchingEnabled", self._frame, state)
-        
-    def make_spectral_reference(self):
-        """Make this image the spectral reference."""
-        self.session.call_action("setSpectralReference", self._frame)
-        
-    @validate(Boolean())
-    def set_spectral_matching(self, state):
-        """Enable or disable spectral matching.
-        
-        Parameters
-        ----------
-        state : boolean
-            The desired spectral matching state.
-        """
-        self.session.call_action("setSpectralMatchingEnabled", self._frame, state)
-
-    # NAVIGATION
-
-    @validate(Evaluate(Number, 0, Attr("depth"), Number.INCLUDE_MIN), Evaluate(Number, 0, Attr("stokes"), Number.INCLUDE_MIN), Boolean())
-    def set_channel_stokes(self, channel=None, stokes=None, recursive=True):
-        """Set the channel and/or Stokes.
-        
-        Parameters
-        ----------
-        channel : {0}
-            The desired channel. Defaults to the current channel.
-        stokes : {1}
-            The desired stokes. Defaults to the current Stokes.
-        recursive : {2}
-            Whether to perform the same change on all spectrally matched images. Defaults to True.
-        """
-        channel = channel or self.get_value("requiredChannel")
-        stokes = stokes or self.get_value("requiredStokes")
-        self.call_action("setChannels", channel, stokes, recursive)
-
-    @validate(Number(), Number())
-    def set_center(self, x, y):
-        """Set the center position.
-        
-        TODO: what are the units?
-        
-        Parameters
-        ----------
-        x : {0}
-            The X position.
-        y : {1}
-            The Y position.
-        """
-        self.call_action("setCenter", x, y)
-        
-    @validate(Number(), Boolean())
-    def set_zoom(self, zoom, absolute=True):
-        """Set the zoom level.
-        
-        TODO: explain this more rigorously.
-        
-        Parameters
-        ----------
-        zoom : {0}
-            The zoom level.
-        absolute : {1}
-            Whether the zoom level should be treated as absolute. By default it is adjusted by a scaling factor.
-        """
-        self.call_action("setZoom", zoom, absolute)
-        
-    # STYLE
-
-    @validate(Constant(Colormap), Boolean())
-    def set_colormap(self, colormap, invert=False):
-        """Set the colormap.
-        
-        Parameters
-        ----------
-        colormap : {0}
-            The colormap.
-        invert : {1}
-            Whether the colormap should be inverted.
-        """
-        self.call_action("renderConfig.setColorMap", colormap)
-        self.call_action("renderConfig.setInverted", invert)
-    
-    # TODO check whether this works as expected
-    @validate(Constant(Scaling), NoneOr(Number()), NoneOr(Number()), NoneOr(Number()), NoneOr(Number()))
-    def set_scaling(self, scaling, alpha=None, gamma=None, min=None, max=None):
-        """Set the colormap scaling.
-        
-        Parameters
-        ----------
-        scaling : {0}
-            The scaling type.
-        alpha : {1}
-            The alpha value (only applicable to ``LOG`` and ``POWER`` scaling types).
-        gamma : {2}
-            The gamma value (only applicable to the ``GAMMA`` scaling type).
-        min : {3}
-            The minimum of the scale. Only used if both *min* and *max* are set.
-        max : {4}
-            The maximum of the scale. Only used if both *min* and *max* are set.
-        """
-        self.call_action("renderConfig.setScaling", scaling)
-        if scaling in (Scaling.LOG, Scaling.POWER) and alpha is not None:
-            self.call_action("renderConfig.setAlpha", alpha)
-        elif scaling == Scaling.GAMMA and gamma is not None:
-            self.call_action("renderConfig.setGamma", gamma)
-        if min is not None and max is not None:
-            self.call_action("renderConfig.setCustomScale", min, max)
-    
-    @validate(Boolean())
-    def set_raster_visible(self, state):
-        """Set the raster image visibility.
-        
-        Parameters
-        ----------
-        state : {0}
-            The desired visibility state.
-        """
-        self.call_action("renderConfig.setVisible", state)
-        
-    def show_raster(self):
-        """Show the raster image."""
-        self.set_raster_visible(True)
-        
-    def hide_raster(self):
-        """Hide the raster image."""
-        self.set_raster_visible(False)
-    
-    # CONTOURS
-    
-    @validate(IterableOf(Number()), Constant(SmoothingMode), Number())
-    def configure_contours(self, levels, smoothing_mode=SmoothingMode.GAUSSIAN_BLUR, smoothing_factor=4):
-        """Configure contours.
-        
-        Parameters
-        ----------
-        levels : {0}
-            The contour levels. This may be a numeric numpy array; e.g. the output of ``arange``.
-        smoothing_mode : {1}
-            The smoothing mode.
-        smoothing_factor : {2}
-            The smoothing factor.
-        """
-        self.call_action("contourConfig.setContourConfiguration", levels, smoothing_mode, smoothing_factor)
-    
-    @validate(NoneOr(Constant(ContourDashMode)), NoneOr(Number()))
-    def set_contour_dash(self, dash_mode=None, thickness=None):
-        """Set the contour dash style.
-        
-        Parameters
-        ----------
-        dash_mode : {0}
-            The dash mode.
-        thickness : {1}
-            The dash thickness.
-        """
-        if dash_mode is not None:
-            self.call_action("contourConfig.setDashMode", dash_mode)
-        if thickness is not None:
-            self.call_action("contourConfig.setThickness", thickness)
-    
-    @validate(Color())
-    def set_contour_color(self, color):
-        """Set the contour color.
-        
-        This automatically disables use of the contour colormap.
-        
-        Parameters
-        ----------
-        color : {0}
-            The color.
-        """
-        self.call_action("contourConfig.setColor", color)
-        self.call_action("contourConfig.setColormapEnabled", False)
-    
-    @validate(Constant(Colormap), NoneOr(Number()), NoneOr(Number()))
-    def set_contour_colormap(self, colormap, bias=None, contrast=None):
-        """Set the contour colormap.
-        
-        This automatically enables use of the contour colormap.
-        
-        Parameters
-        ----------
-        colormap : {0}
-            The colormap.
-        bias : {1}
-            The colormap bias.
-        contrast : {2}
-            The colormap contrast.
-        """
-        self.call_action("contourConfig.setColormap", colormap)
-        self.call_action("contourConfig.setColormapEnabled", True)
-        if bias is not None:
-            self.call_action("contourConfig.setColormapBias", bias)
-        if contrast is not None:
-            self.call_action("contourConfig.setColormapContrast", contrast)
-    
-    def apply_contours(self):
-        """Apply the contour configuration."""
-        self.call_action("applyContours")
-    
-    def clear_contours(self):
-        """Clear the contours."""
-        self.call_action("clearContours", True)
-    
-    @validate(Boolean())
-    def set_contours_visible(self, state):
-        """Set the contour visibility.
-        
-        Parameters
-        ----------
-        state : {0}
-            The desired visibility state.
-        """
-        self.call_action("contourConfig.setVisible", state)
-    
-    def show_contours(self):
-        """Show the contours."""
-        self.set_contours_visible(True)
-        
-    def hide_contours(self):
-        """Hide the contours."""
-        self.set_contours_visible(False)
-    
-    # HISTOGRAM
-    
-    @validate(Boolean())
-    def use_cube_histogram(self, contours=False):
-        """Use the cube histogram.
-        
-        Parameters
-        ----------
-        contours : {0}
-            Apply to the contours. By default this is applied to the raster image.
-        """
-        self.call_action(f"renderConfig.setUseCubeHistogram{'Contours' if contours else ''}", True)
-    
-    @validate(Boolean())
-    def use_channel_histogram(self, contours=False):
-        """Use the channel histogram.
-        
-        Parameters
-        ----------
-        contours : {0}
-            Apply to the contours. By default this is applied to the raster image.
-        """
-        self.call_action(f"renderConfig.setUseCubeHistogram{'Contours' if contours else ''}", False)
-            
-    @validate(Number(0, 100))
-    def set_percentile_rank(self, rank):
-        """Set the percentile rank.
-        
-        Parameters
-        ----------
-        rank : {0}
-            The percentile rank.
-        """
-        self.call_action("renderConfig.setPercentileRank", rank)
-    
-    # CLOSE
-    
-    def close(self):
-        """Close this image."""
-        self.session.call_action("closeFile", self._frame)
